Break out esbuild.CLI public API
