+++ pkg/esbuild/version.go
@@ -1,3 +1,3 @@
-package main
+package esbuild
 
 const esbuildVersion = "0.4.1"
